Extract chatapp runtime inference flow
